requirements.txt
pyside6
pybldc
pyinstaller
uptime
python-can